Move world gen scripts into own file
